ㅕㅕ

+++ src/GEMS_TCO/__init__.py
@@ -164,7 +164,8 @@ class load_data_amarel:
             tmp['Hours_elapsed'] = np.round(tmp['Hours_elapsed']-477700)
 
             tmp = tmp.iloc[ord_mm, :4].to_numpy()  # reorder the data
-            tmp = torch.from_numpy(tmp).float()  # Convert NumPy to Tensor
+            #tmp = torch.from_numpy(tmp).float()  # Convert NumPy to Tensor
+            tmp = torch.from_numpy(tmp).double()  # Convert NumPy to Tensor
             analysis_data_map[key_idx[i]] = tmp
 
         aggregated_data = pd.DataFrame()
@@ -175,7 +176,7 @@ class load_data_amarel:
             aggregated_data = pd.concat((aggregated_data, tmp), axis=0)
 
         aggregated_data = aggregated_data.iloc[:, :4].to_numpy()
-        aggregated_data = torch.from_numpy(aggregated_data).float() 
-
+        #aggregated_data = torch.from_numpy(aggregated_data).float() 
+        aggregated_data = torch.from_numpy(aggregated_data).double()
         return analysis_data_map, aggregated_data
